src/IA/IA_v2/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
opencv-python>=4.6.0
Pillow>=9.0.0
albumentations>=1.3.0
scikit-image>=0.19.0
scikit-learn>=1.1.0
mlflow>=2.0.0
tensorboard>=2.8.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.4.0
tqdm>=4.64.0
psutil>=5.9.0

